+++ sigauth/signature.go
@@ -15,14 +15,12 @@
 package sigauth
 
 import (
-	"bytes"
 	"crypto"
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/base64"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"strconv"
 	"strings"
@@ -91,12 +89,12 @@ func (s *SignatureAuth) CheckSignature(pubKey *rsa.PublicKey, message []byte, si
 }
 
 // SignRequest signs and modifies the provided request with the necessary signature parameters
-func (s *SignatureAuth) SignRequest(r *http.Request) error {
+func (s *SignatureAuth) SignRequest(r *http.Request, body []byte) error {
 	if r == nil {
 		return errors.New("request is nil")
 	}
 
-	digest, length, err := GetRequestDigest(r)
+	digest, length, err := GetRequestDigest(body)
 	if err != nil {
 		return fmt.Errorf("unable to build request digest: %v", err)
 	}
@@ -136,12 +134,12 @@ func (s *SignatureAuth) SignRequest(r *http.Request) error {
 // 	The request must be signed by one of the services in requiredServiceIDs. If nil, any valid signature
 //	from a subscribed service will be accepted
 // 	Returns the service ID of the signing service
-func (s *SignatureAuth) CheckRequestServiceSignature(r *http.Request, requiredServiceIDs []string) (string, error) {
+func (s *SignatureAuth) CheckRequestServiceSignature(r *http.Request, body []byte, requiredServiceIDs []string) (string, error) {
 	if r == nil {
 		return "", errors.New("request is nil")
 	}
 
-	sigString, sigAuthHeader, err := s.checkRequest(r)
+	sigString, sigAuthHeader, err := s.checkRequest(r, body)
 	if err != nil {
 		return "", err
 	}
@@ -160,7 +158,7 @@ func (s *SignatureAuth) CheckRequestServiceSignature(r *http.Request, requiredSe
 
 // CheckRequestSignature validates the signature on the provided request
 // 	The request must be signed by the private key paired with the provided public key
-func (s *SignatureAuth) CheckRequestSignature(r *http.Request, pubKey *rsa.PublicKey) error {
+func (s *SignatureAuth) CheckRequestSignature(r *http.Request, body []byte, pubKey *rsa.PublicKey) error {
 	if r == nil {
 		return errors.New("request is nil")
 	}
@@ -169,7 +167,7 @@ func (s *SignatureAuth) CheckRequestSignature(r *http.Request, pubKey *rsa.Publi
 		return errors.New("public key is nil")
 	}
 
-	sigString, sigAuthHeader, err := s.checkRequest(r)
+	sigString, sigAuthHeader, err := s.checkRequest(r, body)
 	if err != nil {
 		return err
 	}
@@ -182,7 +180,7 @@ func (s *SignatureAuth) CheckRequestSignature(r *http.Request, pubKey *rsa.Publi
 	return nil
 }
 
-func (s *SignatureAuth) checkRequest(r *http.Request) (string, *SignatureAuthHeader, error) {
+func (s *SignatureAuth) checkRequest(r *http.Request, body []byte) (string, *SignatureAuthHeader, error) {
 	authHeader := r.Header.Get("Authorization")
 	if authHeader == "" {
 		return "", nil, errors.New("request missing authorization header")
@@ -190,7 +188,7 @@ func (s *SignatureAuth) checkRequest(r *http.Request) (string, *SignatureAuthHea
 
 	digestHeader := r.Header.Get("Digest")
 
-	digest, _, err := GetRequestDigest(r)
+	digest, _, err := GetRequestDigest(body)
 	if err != nil {
 		return "", nil, fmt.Errorf("unable to build request digest: %v", err)
 	}
@@ -264,21 +262,10 @@ func GetRequestLine(r *http.Request) string {
 }
 
 // GetRequestDigest returns the SHA256 digest and length of the provided request body
-func GetRequestDigest(r *http.Request) (string, int, error) {
-	if r == nil {
-		return "", 0, errors.New("request is nil")
-	}
-	if r.Body == nil {
+func GetRequestDigest(body []byte) (string, int, error) {
+	if len(body) == 0 {
 		return "", 0, nil
 	}
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return "", 0, fmt.Errorf("error reading request body: %v", err)
-	}
-	r.Body.Close()
-
-	r.Body = ioutil.NopCloser(bytes.NewReader(body))
 
 	hash, err := authutils.HashSha256(body)
 	if err != nil {